+++ lib/features/working_directory/presentation/bloc/working_directory_bloc.dart
@@ -28,12 +28,97 @@ class WorkingDirectoryBloc extends Bloc<WorkingDirectoryEvent, WorkingDirectoryS
       final commandResult = await gitService.runGit(GitCommands.statusPorcelain, repositoryPath);
 
       final files = gitService.parseGitStatusPorcelain(commandResult);
+      final commitsToPush = await gitService.getCommitsAheadCount(repositoryPath);
 
       emit(
         state.copyWith(
+          commitsToPush: commitsToPush,
           files: files,
         ),
       );
+    });
+
+    on<PushCommits>((event, emit) async {
+      final repositoryPath = sharedPreferencesService.getString(SharedPreferencesKeys.repositoryPath) ?? "";
+
+      if (repositoryPath.isEmpty) return;
+
+      try {
+        emit(
+          state.copyWith(
+            status: WorkingDirectoryBlocStatus.loading,
+          ),
+        );
+
+        await gitService.convertRemoteToSsh(repositoryPath);
+        await gitService.runGit(
+          GitCommands.gitPush,
+          repositoryPath,
+        );
+
+        add(GetRepositoryStatus());
+      } catch (e) {
+        emit(
+          state.copyWith(
+            status: WorkingDirectoryBlocStatus.error,
+            errorMessage: e.toString(),
+          ),
+        );
+      }
+    });
+
+    on<UpdateWorkingDirectoryStatus>((event, emit) {
+      emit(state.copyWith(status: event.status));
+    });
+
+    on<AddCommit>((event, emit) async {
+      final repositoryPath = sharedPreferencesService.getString(SharedPreferencesKeys.repositoryPath) ?? "";
+
+      if (repositoryPath.isEmpty) return;
+
+      final List<String> args = [
+        ...GitCommands.gitCommit,
+        '-m',
+        event.summary,
+      ];
+
+      final description = event.description?.trim();
+      if (description != null && description.isNotEmpty) {
+        args.addAll(['-m', description]);
+      }
+
+      await gitService.runGit(args, repositoryPath);
+
+      add(GetRepositoryStatus());
+    });
+
+    on<ToggleFileStaging>((event, emit) async {
+      final repositoryPath = sharedPreferencesService.getString(SharedPreferencesKeys.repositoryPath) ?? "";
+
+      if (repositoryPath.isEmpty) return;
+
+      try {
+        if (event.stage) {
+          await gitService.runGit(
+            [...GitCommands.gitAdd, event.file.path],
+            repositoryPath,
+          );
+        } else {
+          await gitService.runGit(
+            [...GitCommands.gitRestoreStaged, event.file.path],
+            repositoryPath,
+          );
+        }
+
+        add(GetRepositoryStatus());
+      } catch (e) {
+        emit(
+          state.copyWith(
+            status: WorkingDirectoryBlocStatus.error,
+            errorMessage: e.toString(),
+          ),
+        );
+      }
     });
   }
 }