+++ lib/features/working_directory/presentation/bloc/working_directory_event.dart
@@ -5,3 +5,35 @@ sealed class WorkingDirectoryEvent {}
 class GetRepositoryStatus extends WorkingDirectoryEvent {
   GetRepositoryStatus();
 }
+
+class AddCommit extends WorkingDirectoryEvent {
+  final String summary;
+  final String? description;
+
+  AddCommit({
+    required this.summary,
+    this.description,
+  });
+}
+
+class ToggleFileStaging extends WorkingDirectoryEvent {
+  final GitFileEntity file;
+  final bool stage;
+
+  ToggleFileStaging({
+    required this.file,
+    required this.stage,
+  });
+}
+
+class PushCommits extends WorkingDirectoryEvent {
+  PushCommits();
+}
+
+class UpdateWorkingDirectoryStatus extends WorkingDirectoryEvent {
+  final WorkingDirectoryBlocStatus status;
+
+  UpdateWorkingDirectoryStatus({
+    required this.status,
+  });
+}