+++ lib/features/working_directory/presentation/bloc/working_directory_state.dart
@@ -12,10 +12,12 @@ class WorkingDirectoryState with WorkingDirectoryStateMappable {
   final WorkingDirectoryBlocStatus status;
   final List<GitFileEntity> files;
   final String errorMessage;
+  final int commitsToPush;
 
   const WorkingDirectoryState({
     this.status = WorkingDirectoryBlocStatus.initial,
     this.files = const [],
     this.errorMessage = "",
+    this.commitsToPush = 0,
   });
 }