+++ lib/features/working_directory/presentation/ui/working_directory_files_list.dart
@@ -1,17 +1,26 @@
 import 'package:flutter/material.dart';
 import 'package:open_git/shared/domain/entities/git_file_entity.dart';
 import 'package:open_git/features/working_directory/presentation/ui/working_directory_item.dart';
+import 'package:open_git/shared/presentation/widgets/commit_message_textfield.dart';
 
 class WorkingDirectoryFilesList extends StatelessWidget {
   final List<GitFileEntity> files;
-  final ValueChanged<GitFileEntity> onCheckboxToggled;
+  final Function({bool? isChecked, required GitFileEntity file}) onCheckboxToggled;
   final Function(GitFileEntity file) onFileSelected;
+  final bool hasStagedFiles;
+  final void Function({
+    required String summary,
+    required String description,
+  })
+  onCommitPressed;
 
   const WorkingDirectoryFilesList({
     super.key,
     required this.files,
     required this.onCheckboxToggled,
     required this.onFileSelected,
+    required this.hasStagedFiles,
+    required this.onCommitPressed,
   });
 
   @override
@@ -32,8 +41,8 @@ class WorkingDirectoryFilesList extends StatelessWidget {
               final file = files[index];
               return WorkingDirectoryItem(
                 file: file,
-                onCheckboxToggled: (_) {
-                  onCheckboxToggled(file);
+                onCheckboxToggled: (isChecked) {
+                  onCheckboxToggled(isChecked: isChecked, file: file);
                 },
                 onSelected: (file) {
                   onFileSelected(file);
@@ -41,6 +50,10 @@ class WorkingDirectoryFilesList extends StatelessWidget {
               );
             },
           ),
+        ),
+        CommitMessageTextfield(
+          onCommitPressed: onCommitPressed,
+          hasStagedFiles: hasStagedFiles,
         ),
       ],
     );